Replaced: Transparency and Behaviour with Power
The new Power system is slightly more complex but is a lot more flexible and powerful and enables new use cases like an obligation (the receiver of a value is obliged to undertake a certain action or compilation fails). Thanks to the fact that powers are binary (have or don't) implementation isstill simpler if not simpler. The only not so nice thing is that certain powers are recursive (the one associated with the previous behaviour) where others are not (the one associated with the previous transparency).

The biggest + point is that the structure here is reversed meaning that a lot of generic power declaration fall away and per default the developer is forced to work with a powerless type and must explicitly opt into power, preventing missing declarations out of a mistake as the compiler will catch it now.